chore(middlewares): consolidate to plural directory convention
